requirements/base.txt
django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-filter==22.1
django-autoslug==1.9.8
django-countries==7.5.1
django-phonenumber-field==7.0.2
phonenumbers==8.13.7
colorlog==6.8.0
drf-yasg==1.21.5
Pillow==9.4.0
argon2-cffi==21.3.0
pytz==2022.7.1
django-taggit==3.1.0
redis==4.5.1
celery==5.2.7
flower==1.2.0
django-celery-email==3.0.0
django-allauth==0.52.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==3.0.0